Fixed ETF furnace short overflow bug (#113)

+++ src/main/java/ganymedes01/etfuturum/tileentities/TileEntityBlastFurnace.java
@@ -129,7 +129,7 @@ public class TileEntityBlastFurnace extends TileEntity implements ISidedInventor
 		this.furnaceItemStacks = new ItemStack[this.getSizeInventory()];
 		Utils.loadItemStacksFromNBT(nbttaglist, this.furnaceItemStacks);
 
-		this.furnaceBurnTime = compound.getShort("BurnTime");
+		this.furnaceBurnTime = compound.getInteger("BurnTime");
 		this.furnaceCookTime = compound.getShort("CookTime");
 		this.currentItemBurnTime = getItemBurnTime(this.furnaceItemStacks[1]);
 
@@ -141,7 +141,7 @@ public class TileEntityBlastFurnace extends TileEntity implements ISidedInventor
 	@Override
 	public void writeToNBT(NBTTagCompound compound) {
 		super.writeToNBT(compound);
-		compound.setShort("BurnTime", (short) this.furnaceBurnTime);
+		compound.setInteger("BurnTime", this.furnaceBurnTime);
 		compound.setShort("CookTime", (short) this.furnaceCookTime);
 
 		compound.setTag("Items", Utils.writeItemStacksToNBT(this.furnaceItemStacks));
